requirements.txt
# Core Framework
streamlit==1.29.0
streamlit-authenticator==0.2.3
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4.post3
streamlit-lottie==0.0.5

# Data & Analysis
pandas==2.0.3
numpy==1.24.3
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.13.0
altair==5.2.0

# AI & NLP
textblob==0.17.1
nltk==3.8.1
google-generativeai==0.3.2
transformers==4.35.2
torch==2.1.0
spacy==3.7.2

# Web & APIs
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.0
beautifulsoup4==4.12.2
selenium==4.15.2

# Google Cloud Services
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-vision==3.4.4
google-cloud-language==2.11.1
google-cloud-translate==3.11.1
google-cloud-monitoring==2.16.0
google-auth==2.23.4

# Firebase
firebase-admin==6.2.0
pyrebase4==4.7.1

# Security
cryptography==41.0.7
bcrypt==4.1.1
python-jose==3.3.0

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
scikit-image==0.22.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pathlib==1.0.1
faker==20.1.0

# Development
pytest==7.4.3
black==23.10.1
